% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmode_tikz.R
\name{str.figure}
\alias{str.figure}
\title{Export a latex figure from a string input}
\usage{
\method{str}{figure}(
  figure.string,
  fig.float = "t",
  fig.caption = "My figure",
  fig.label = "fig:my_figure"
)
}
\arguments{
\item{figure.string}{The string to use inside the figure block}

\item{fig.float}{Latex figure float modifiers as a string (e.g. "htbp", "!h")}

\item{fig.caption}{Latex figure caption}

\item{fig.label}{Latex figure label for referencing}
}
\value{
The latex figure block as a string
}
\description{
Create a latex figure where the input to the function is a string.
This sets up the surrounding string.
}
