% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmode_tikz.R
\name{tex.figure}
\alias{tex.figure}
\title{Export a latex figure from a tex file}
\usage{
tex.figure(
  inpath,
  fig.float = "t",
  fig.caption = "My figure",
  fig.label = "fig:my_figure",
  is.subfigure = FALSE,
  subfigure.scale = 1
)
}
\arguments{
\item{inpath}{Filepath to the location of the tex file}

\item{fig.float}{Latex figure float modifiers as a string (e.g. "htbp", "!h")}

\item{fig.caption}{Latex figure caption}

\item{fig.label}{Latex figure label for referencing}

\item{is.subfigure}{Boolean value to create a subfigure instead of normal figure}

\item{subfigure.scale}{Set width of subfigure if is.subfigure is TRUE}
}
\value{
A latex figure as a string
}
\description{
Create a latex figure using a tikz .tex file as the input
to the figure. This is useful for placing tikz figures into org-mode
files.
}
